Refactored utils into lib folder
