+++ lib/math-utils.ts
